human/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
django-cors-headers==3.7.0
django-environ==0.4.5
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
phonenumbers==8.12.24
Pygments==2.9.0
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
